% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sample_estimator.R
\name{pooled_smoothed}
\alias{pooled_smoothed}
\title{Logconcave estimator of the location-shift in the two sample model.}
\usage{
pooled_smoothed(dat, eta = 1e-04)
}
\arguments{
\item{dat}{A list with two components: x and y, each being vector of possibly different lengths; represents the data.}

\item{eta}{A fraction between 0 and 1/2. Corresponds to the truncation level of the one step estimator.
The default is 0.0001.}
}
\value{
A vector of length two.
\itemize{
\item \code{estimate:}  The estimated value of \eqn{\Delta}.
\item \code{FI:} The estimated Fisher information for estimating \eqn{\Delta}.
}
}
\description{
Suppose m univariate observations \eqn{X_1, \ldots, X_m} are sampled from a density \eqn{g(x-\mu)},
and n univariate observations \eqn{Y_1, \ldots, Y_n} are sampled from the density \eqn{g(x-\mu-\Delta)}, where
\eqn{g} is an unknown log-concave density. This function computes
a one step estimator to estimte \eqn{\Delta}. This estimator relies on
the smoothed log-concave MLE estimator from the package \code{\link{logcondens}} to estimate \eqn{g}, and  is root-n consistent
for \eqn{\Delta} provided \eqn{g} is log-concave.
}
\details{
\code{eta:} If eta is zero, the function computes the one step estimator  without any
truncation. See Saha et al. (2023) for more details.
}
\examples{
x <- rlogis(100); y <- rlogis(150) + 0.1; 
pooled_smoothed(list(x=x, y=y), eta = 0.0001)
}
\references{
Saha R., Dey P., Laha N. (2023). \emph{Revisiting the two-sample location model
with log-concavity assumption}. submitted.
}
\author{
\href{https://www.nilanjana91.de/}{Nilanjana Laha}
(maintainer), \email{nlaha@tamu.edu}.\cr
Ridhiman Saha, \email{riddhimansaha@fas.harvard.edu}
}
