% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSL.logconcave-package.R
\docType{package}
\name{TSL.logconcave-package}
\alias{TSL.logconcave}
\alias{TSL.logconcave-package}
\title{TSL.logconcave: Estimates the location shift in a two sample location estimation problem}
\description{
Estimates the location shift in a two sample location shift problem with the additional assumption that the underlying densities are log-concave.
}
\author{
\strong{Maintainer}: Nilanjana Laha \email{nlaha@tamu.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
